m4t_server/requirements.txt
grpcio==1.59.0
protobuf==4.24.4
scipy==1.11.3
TTS==0.19.0
